refactor: Improve code readability
